Rename bounty->bountues
