+++ components/bounties/BountyCard.tsx
