+++ components/common/page-layout/Sidebar.tsx
@@ -262,7 +262,7 @@ export default function Sidebar ({ closeSidebar, favorites }: SidebarProps) {
               />
             </Box>
           </Box>
-          <Link href={`/${space.domain}/bounty`}>
+          <Link href={`/${space.domain}/bounties`}>
             <BountyLabel>
               <Typography sx={{ color: greyColor2, fontSize: 12, letterSpacing: '0.03em', fontWeight: 600, px: 2, mb: 0.5 }}>
                 {/* //xtungvo TODO: styling this later */}
